requirements.txt
ascii-magic==2.3.0
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
decorator==5.1.1
executing==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
kaitaistruct==0.10
lxml==4.9.3
matplotlib-inline==0.1.6
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
Pillow==10.1.0
prettytable==3.9.0
prompt-toolkit==3.0.41
pure-eval==0.2.2
pyasn1==0.5.1
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.19.0
pydivert==2.1.0
Pygments==2.17.2
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
pystyle==2.9
pytesseract==0.3.10
requests==2.31.0
requests-toolbelt==1.0.0
selenium==4.14.0
selenium-wire==5.1.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
streamlink==6.3.1
tls-client==0.2.2
traitlets==5.14.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
undetected-chromedriver==3.5.4
urllib3==2.0.7
wcwidth==0.2.8
websocket-client==1.6.4
websockets==12.0
wsproto==1.2.0
zstandard==0.22.0
